feat(core): glamorous tiny
