+++ package.json
@@ -28,7 +28,7 @@
   "author": "Kent C. Dodds <kent@doddsfamily.us> (http://kentcdodds.com/)",
   "license": "MIT",
   "dependencies": {
-    "brcast": "^1.1.6",
+    "brcast": "^2.0.0",
     "fast-memoize": "^2.2.7",
     "html-tag-names": "^1.1.1",
     "react-html-attributes": "^1.2.1",
